requirements.txt
dotenv
diffusers
torch
transformers
accelerate
tqdm
numpy
pillow
opencv-python
ultralytics
matplotlib
